+++ Rdmp.Core/Curation/Data/Datasets/HDR/HDRDatasetPatch.cs
@@ -9,23 +9,7 @@ using System.Threading.Tasks;
 
 namespace Rdmp.Core.Curation.Data.Datasets.HDR
 {
-    public class PatchMetadata
-    {
 
-        public string schemaModel { get; set; }
-        public string schemaVersion { get; set; }
-        public PatchSubMetadata metadata { get; set; }
-
-
-        public PatchMetadata() { }
-        public PatchMetadata(HDRDatasetItems.Metadata existingMetadata)
-        {
-
-            metadata = existingMetadata.metadata != null ? new PatchSubMetadata(existingMetadata.metadata) : null;
-            schemaModel = "HDRUK";
-            schemaVersion = "3.0.0";
-        }
-    }
 
     public class PatchSubMetadata
     {
